requirements.txt
Django==5.0.1
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.3
Pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0